pyproject.toml
[tool.poetry]
name = "xgress"
version = "1.0.0"
description = ""
authors = ["Roman Hlushko <roman.glushko.m@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
aiofiles = "^23.1.0"
fastapi = "^0.95.0"
typer = "^0.7.0"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
